requirements.txt
# InsightBoard Dependencies

# Core Data Processing
pandas==2.1.3
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Streamlit Dashboard
streamlit==1.29.0

# Configuration
pyyaml==6.0.1

# Database (if needed)
sqlalchemy==2.0.23

# Statistics
scipy==1.11.4

# Utilities
python-dateutil==2.8.2